Implement WalkContext and Rule.applyWithFunction
